fix youtube pagination bug
We weren't setting the new pageToken on each pagination request because
our expression was evaluating to false due to a bad condition. We now
ensure a request page exists and does not equal the first request page
that gets set to 1 by default on each new subsequent pagination request.

test plan:
• verify youtube search still fetches first 25 items
• verify "load more results" fetches a new set of 25 items